requirements.txt
stable-baselines3==2.0.0
cloudpickle==2.2.1
filterpy==1.4.5 
hurst
finta==1.3
graphviz
gymnasium==0.28.1
jupyter
keras==2.13.1
neat-python==0.92
numpy==1.24.3
openpyxl==3.1.2
pandas==1.5.3
pip==23.2.1
quantstats
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
setuptools==68.0.0
shimmy==1.2.1
six==1.16.0
statsmodels
tqdm
wheel